+++ libs/gl-client-py/CHANGELOG.md
@@ -5,3 +5,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
+### Changed
+
+- Signer version updated from `v25.12` to `v26.06` (VLS 0.14.0 / CLN v26.06.1)
+- Remove the use of `decodepay`, deprecated in v24.11.